Use double

+++ SaveUtils.hpp
@@ -5,7 +5,7 @@
 
 namespace SaveUtils {
 
-std::string rgbToPpm(glm::vec3** rgbImage, const int width, const int height);
+std::string rgbToPpm(glm::dvec3** rgbImage, const int width, const int height);
 void writePpm(const char* outPath, std::string ppmString);
 
 } // namespace SaveUtils
