+++ src/Transform.cpp
@@ -7,8 +7,8 @@
 
 namespace Transform {
 
-glm::mat4 translate(const float x, const float y, const float z) {
-    return glm::transpose(glm::mat4(
+glm::dmat4 translate(const double x, const double y, const double z) {
+    return glm::transpose(glm::dmat4(
         1, 0, 0, x,
         0, 1, 0, y, 
         0, 0, 1, z,
@@ -16,7 +16,7 @@ glm::mat4 translate(const float x, const float y, const float z) {
     ));
 }
 
-glm::mat4 scale(const float x, const float y, const float z) {
+glm::dmat4 scale(const double x, const double y, const double z) {
     return glm::transpose(glm::mat4(
         x, 0, 0, 0,
         0, y, 0, 0,
@@ -26,19 +26,19 @@ glm::mat4 scale(const float x, const float y, const float z) {
 }
 
 // Rodriguez rotation formula
-glm::mat4 rotate(const glm::vec3 axis, const float theta) {
-    const glm::mat3 I = glm::mat3(1.0f);
-    const glm::mat3 crossMat = glm::transpose(glm::mat3(
+glm::dmat4 rotate(const glm::dvec3 axis, const double theta) {
+    const glm::dmat3 I = glm::mat3(1.0);
+    const glm::dmat3 crossMat = glm::transpose(glm::dmat3(
         0, -axis[2], axis[1],
         axis[2], 0, -axis[0],
         -axis[1], axis[0], 0
     ));
-    const glm::mat3 dotMat = glm::transpose(glm::mat3(
+    const glm::dmat3 dotMat = glm::transpose(glm::dmat3(
         axis[0]*axis[0], axis[0]*axis[1], axis[0]*axis[2],
         axis[1]*axis[0], axis[1]*axis[1], axis[1]*axis[2],
         axis[2]*axis[0], axis[2]*axis[1], axis[2]*axis[2]
     ));
-    return (I*cos(theta)) + (crossMat*sin(theta)) + (dotMat * (1.0f - cos(theta)));
+    return (I*cos(theta)) + (crossMat*sin(theta)) + (dotMat * (1.0 - cos(theta)));
 }
